ids are changed to int64 and telegram id is added to user context

+++ internal/repository/user.go
@@ -15,14 +15,14 @@ func NewUserPostgres(db *sqlx.DB) *UserPostgres {
 	return &UserPostgres{db: db}
 }
 
-func (r *UserPostgres) GetUserInfo(id int) (core.User, error) {
+func (r *UserPostgres) GetUserInfo(id int64) (core.User, error) {
 	var destUser core.User
 	query := fmt.Sprintf(`SELECT id, first_name, last_name, username, rating, num_people_rated, tg_id  FROM %s WHERE id=$1`, usersTable)
 	err := r.db.Get(&destUser, query, id)
 	return destUser, err
 }
 
-func (r *UserPostgres) JoinTrip(userId, tripId int) error {
+func (r *UserPostgres) JoinTrip(userId, tripId int64) error {
 	query := fmt.Sprintf(`INSERT INTO %s (user_id, trip_id) VALUES ($1, $2)`, usersTripsTable)
 	_, err := r.db.Exec(query, userId, tripId)
 	return err
