+++ internal/transport/http/v1/user.go
@@ -38,7 +38,7 @@ func (h *Handler) redirectReqToBot(c *gin.Context) {
 		return
 	}
 
-	trip, err := h.services.Trip.GetById(trip_id)
+	trip, err := h.services.Trip.GetById(int64(trip_id))
 	if err != nil {
 		response.NewErrorResponse(c, http.StatusInternalServerError, err.Error())
 		return
@@ -46,7 +46,7 @@ func (h *Handler) redirectReqToBot(c *gin.Context) {
 	var redirectReq = joinRequest{
 		UserId:      uctx.UserId,
 		AdminId:     trip.AdminId,
-		TripId:      trip_id,
+		TripId:      int64(trip_id),
 		SecretToken: os.Getenv("BACKEND_SECRET_TOKEN"),
 		TripName:    getTripName(trip.FromPoint, trip.ToPoint, trip.ChosenTimestamp),
 	}
@@ -83,9 +83,9 @@ func (h *Handler) getResFromBot(c *gin.Context) { // TODO: webhook
 }
 
 type joinRequest struct {
-	AdminId     int    `json:"trip_admin_id"`
-	TripId      int    `json:"trip_id" binding:"required"`
-	UserId      int    `json:"id_of_person_asking_to_join" binding:"required"`
+	AdminId     int64  `json:"trip_admin_id"`
+	TripId      int64  `json:"trip_id" binding:"required"`
+	UserId      int64  `json:"id_of_person_asking_to_join" binding:"required"`
 	SecretToken string `json:"secret_token" binding:"required"`
 	Accepted    bool   `json:"accepted"`
 	TripName    string `json:"trip_name"`
